Deal with subsequent exec()s (not part of the fork-exec pattern)
Initially unsure if a subsequent exec() should _replace_ the original
executable/argv/env details of the ProcessTrace instance, but after some
thinking, I believe it's better to preserve the original process' details
as the user of the ProcessTrace instance is most likely holding the POV of
a parent process expecting to find an instance whose argv matches the one
that was originally given to strace_helper.run_trace().

Therefore, subsequent exec()s in a ProcessTrace instance will only be
registered as simple read()s of the executable passed to exec(). All
subsequent activity in the exec()ed process will be registered as part
of the same ProcessTrace instance.

Example: Running "sh -c 'cat /dev/null'" involves first (fork()ing and)
exec()ing sh, which then ends up exec()ing (without first fork()ing) cat
- which finally ends up reading /dev/null.

Instead of registering the 'cat' "process" in a separate ProcessTrace
instance, or - alternatively - mutating the original ProcessTrace
instance into adopting the 'cat' process' details, we instead "pretend"
(not really) that the 'cat' is merely a continuation of the 'sh'
process, and register the exec as a simple read of the cat executable.
Any activity performed by 'cat' (reading /dev/null in this case) is
then merely registered as yet another activity on the original (sh)
ProcessTrace.

In short, the ProcessTrace instance chooses to strictly trac the
behavior of a PID, even when subsequent exec()s causes that PID to take
on the "personality" of several different logical processes.